Resolve conflicts after Distance merge.

+++ server/src/main/java/com/tripco/t00/planner/TFFIResponse.java
@@ -1,6 +1,6 @@
 package com.tripco.t00.planner;
 
 public abstract class TFFIResponse {
-  protected short version;
+  protected Integer version;
   protected String type;
 }
